HandwritingRecognition/mnist_desktop_project/requirements.txt
tqdm
numpy
opencv-contrib-python
pyserial
pygame
